Initial support for #278
As part of supporting BeanParam I added a new ParameterMetadata collection to MethodMetadata. This collection is intended to store additional metadata about each parameter. I feel like in addition to the Transformer which is the sole member of this clas, Expander could also be placed here, but I opted not to break the API.

For now, Transformer is an instance instead of a class like the Expander metadata. This allows me to configure it as I am assembling the contract. Instead of stashing the BeanParam metadata in some sort of contract context, or an additional field(s) on the metadata, only for it to be picked up and reflective constructed later.

This is still work in progress, but is functional.